+++ src/js/Bounds.d.ts
@@ -3,8 +3,10 @@
  * @date     2020-05-11
  * @modified 2020-10-30 Added the static computeFromVertices function.
  * @modified 2020-11-19 Set min, max, width and height to private.
- * @version  1.1.1
+ * @modified 2021-02-02 Added the `toPolygon` method.
+ * @version  1.2.0
  **/
+import { Polygon } from "./Polygon";
 import { XYCoords, IBounds, XYDimension } from "./interfaces";
 import { Vertex } from "./Vertex";
 /**
@@ -52,6 +54,15 @@ export declare class Bounds implements IBounds, XYDimension {
      * @param {XYCoords} max - The max values (x,y) as a XYCoords tuple.
      **/
     constructor(min: XYCoords, max: XYCoords);
+    /**
+     * Convert this rectangular bounding box to a polygon with four vertices.
+     *
+     * @method toPolygon
+     * @instance
+     * @memberof Bounds
+     * @return {Polygon} This bound rectangle as a polygon.
+     */
+    toPolygon(): Polygon;
     /**
      * Compute the minimal bounding box for a given set of vertices.
      *