+++ src/js/Bounds.js
@@ -4,10 +4,12 @@
  * @date     2020-05-11
  * @modified 2020-10-30 Added the static computeFromVertices function.
  * @modified 2020-11-19 Set min, max, width and height to private.
- * @version  1.1.1
+ * @modified 2021-02-02 Added the `toPolygon` method.
+ * @version  1.2.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Bounds = void 0;
+var Polygon_1 = require("./Polygon");
 var Vertex_1 = require("./Vertex");
 /**
  * @classdesc A bounds class with min and max values. Implementing IBounds.
@@ -31,6 +33,23 @@ var Bounds = /** @class */ (function () {
         this.width = max.x - min.x;
         this.height = max.y - min.y;
     }
+    ;
+    /**
+     * Convert this rectangular bounding box to a polygon with four vertices.
+     *
+     * @method toPolygon
+     * @instance
+     * @memberof Bounds
+     * @return {Polygon} This bound rectangle as a polygon.
+     */
+    Bounds.prototype.toPolygon = function () {
+        return new Polygon_1.Polygon([
+            new Vertex_1.Vertex(this.min),
+            new Vertex_1.Vertex(this.max.x, this.min.y),
+            new Vertex_1.Vertex(this.max),
+            new Vertex_1.Vertex(this.min.x, this.max.y)
+        ], false);
+    };
     ;
     /**
      * Compute the minimal bounding box for a given set of vertices.