+++ src/js/utils/algorithms/sutherlandHodgman.js
@@ -5,7 +5,7 @@
  * Original version:
  *    https://rosettacode.org/wiki/Sutherland-Hodgman_polygon_clipping#JavaScript
  *
- * @author  Ikaros Kappler (ported to TypeScript with {x,y} vertices.
+ * @author  Ikaros Kappler (ported to TypeScript with {x,y} vertices).
  * @date    2021-01-29
  * @version 1.0.0
  */
@@ -28,24 +28,28 @@ var intersection = function (cp1, cp2, s, e) {
     var dc = {
         x: cp1.x - cp2.x,
         y: cp1.y - cp2.y
-    }, dp = { x: s.x - e.x,
+    };
+    var dp = {
+        x: s.x - e.x,
         y: s.y - e.y
-    }, n1 = cp1.x * cp2.y - cp1.y * cp2.x, n2 = s.x * e.y - s.y * e.x, n3 = 1.0 / (dc.x * dp.y - dc.y * dp.x);
+    };
+    var n1 = cp1.x * cp2.y - cp1.y * cp2.x;
+    var n2 = s.x * e.y - s.y * e.x;
+    var n3 = 1.0 / (dc.x * dp.y - dc.y * dp.x);
     return { x: (n1 * dp.x - n2 * dc.x) * n3,
         y: (n1 * dp.y - n2 * dc.y) * n3
     };
 };
 /**
- * @param {Array<XYCoords>} subjectPolygon
+ * @param {Array<XYCoords>} subjectPolygon - Can be any polygon.
  * @param {Array<XYCoords>} clipPolygon - Must be convex.
  */
 var sutherlandHodgman = function (subjectPolygon, clipPolygon) {
-    var cp1 = clipPolygon[clipPolygon.length - 1]; // = { x : 0, y : 0 };
-    var cp2; //  = { x : 0, y : 0 };
-    var s; // = { x : 0, y : 0 };
-    var e; //  = { x : 0, y : 0 };
+    var cp1 = clipPolygon[clipPolygon.length - 1];
+    var cp2;
+    var s;
+    var e;
     var outputList = subjectPolygon;
-    // cp1 = clipPolygon[clipPolygon.length-1];
     for (var j in clipPolygon) {
         cp2 = clipPolygon[j];
         var inputList = outputList;