+++ changelog.md
@@ -1,11 +1,14 @@
 
 
 # Changelog
+* 2021-02-02
+  * Added the `Bounds.toPolygon` method.
 * 2021-01-29
   * Added the `Polygon.signedArea` function (was global function in the demos before).
   * Added the `Polygon.isClockwise` function.
   * Added the `Polygon.area` function.
   * Changed the param type for `containsVert` from Vertex to XYCoords.
+  * Added an implementation of the Sutherland-Hodgman polygon clipping algorithm.
 * 2021-01-27 [v1.11.0]
   * Added the new svg renderer as integrated helper.
 * 2021-01-26